simplify cmd + logs
